Move lookup functionality into a trait.
This provides us with two things: First, it allows us to easily add
lookup functionality to other types than the base Config. This is
important in the use case of partially parsing a configuration in
different sections of the code. For example, we may look up a section
in one place and then find values within that section in another.
Second, since all of the lookup_x functions are based on lookup, we
can write them all in one place and implementors of the Lookup trait
only need to write a lookup function instead of needing to reimplement
all of the individual lookup helpers.

Additionally, since lookup() returns a value, implement Lookup on
the Value type. This means that

    cfg.lookup_integer32("section.value")

is equivalent to

    cfg.lookup("section").unwrap().lookup_integer32("value")

assuming that the configuration actually has the specified value.

Finally, added an as_value() method to the Config type. While
possibly not important right this moment, it allows us to treat a
configuration as just another value, since under the hood that's all
it is. This is much in the same as the as_raw_fd() functionality for
a std::fs::File, granting access to the underlying structure.